+++ packages/shared/common/jest.config.js
@@ -4,5 +4,4 @@ module.exports = {
   testEnvironment: 'node',
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
   collectCoverageFrom: ['src/**/*.ts'],
-  setupFilesAfterEnv: ['@launchdarkly/private-js-mocks/setup'],
 };