+++ packages/shared/sdk-client/src/LDClientImpl.events.test.ts
@@ -6,10 +6,9 @@ import {
   LDContext,
   subsystem,
 } from '@launchdarkly/js-sdk-common';
-import { InputCustomEvent, InputIdentifyEvent } from '@launchdarkly/js-sdk-common/dist/internal';
 import {
-  basicPlatform,
-  logger,
+  createBasicPlatform,
+  createLogger,
   MockEventProcessor,
   setupMockStreamingProcessor,
 } from '@launchdarkly/private-js-mocks';
@@ -17,6 +16,17 @@ import {
 import * as mockResponseJson from './evaluation/mockResponse.json';
 import LDClientImpl from './LDClientImpl';
 import { Flags } from './types';
+
+type InputCustomEvent = internal.InputCustomEvent;
+type InputIdentifyEvent = internal.InputIdentifyEvent;
+
+let mockPlatform: ReturnType<typeof createBasicPlatform>;
+let logger: ReturnType<typeof createLogger>;
+
+beforeEach(() => {
+  mockPlatform = createBasicPlatform();
+  logger = createLogger();
+});
 
 jest.mock('@launchdarkly/js-sdk-common', () => {
   const actual = jest.requireActual('@launchdarkly/js-sdk-common');
@@ -57,9 +67,9 @@ describe('sdk-client object', () => {
       }),
     );
     setupMockStreamingProcessor(false, defaultPutResponse);
-    basicPlatform.crypto.randomUUID.mockReturnValue('random1');
+    mockPlatform.crypto.randomUUID.mockReturnValue('random1');
 
-    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, basicPlatform, {
+    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, mockPlatform, {
       logger,
     });
     jest