+++ packages/shared/sdk-client/src/LDClientImpl.test.ts
@@ -1,8 +1,7 @@
-import { AutoEnvAttributes, clone, LDContext } from '@launchdarkly/js-sdk-common';
+import { AutoEnvAttributes, clone, Hasher, LDContext } from '@launchdarkly/js-sdk-common';
 import {
-  basicPlatform,
-  hasher,
-  logger,
+  createBasicPlatform,
+  createLogger,
   MockStreamingProcessor,
   setupMockStreamingProcessor,
 } from '@launchdarkly/private-js-mocks';
@@ -42,17 +41,25 @@ const autoEnv = {
     os: { name: 'An OS', version: '1.0.1', family: 'orange' },
   },
 };
-let ldc: LDClientImpl;
-let defaultPutResponse: Flags;
-
 describe('sdk-client object', () => {
+  let ldc: LDClientImpl;
+  let defaultPutResponse: Flags;
+  let mockPlatform: ReturnType<typeof createBasicPlatform>;
+  let logger: ReturnType<typeof createLogger>;
+
   beforeEach(() => {
+    mockPlatform = createBasicPlatform();
+    logger = createLogger();
     defaultPutResponse = clone<Flags>(mockResponseJson);
     setupMockStreamingProcessor(false, defaultPutResponse);
-    basicPlatform.crypto.randomUUID.mockReturnValue('random1');
-    hasher.digest.mockReturnValue('digested1');
+    mockPlatform.crypto.randomUUID.mockReturnValue('random1');
+    const hasher = {
+      update: jest.fn((): Hasher => hasher),
+      digest: jest.fn(() => 'digested1'),
+    };
+    mockPlatform.crypto.createHash.mockReturnValue(hasher);
 
-    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, basicPlatform, {
+    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, mockPlatform, {
       logger,
       sendEvents: false,
     });
@@ -61,7 +68,8 @@ describe('sdk-client object', () => {
       .mockReturnValue('/stream/path');
   });
 
-  afterEach(() => {
+  afterEach(async () => {
+    await ldc.close();
     jest.resetAllMocks();
   });
 
@@ -84,6 +92,8 @@ describe('sdk-client object', () => {
   test('identify success', async () => {
     defaultPutResponse['dev-test-flag'].value = false;
     const carContext: LDContext = { kind: 'car', key: 'test-car' };
+
+    mockPlatform.crypto.randomUUID.mockReturnValue('random1');
 
     await ldc.identify(carContext);
     const c = ldc.getContext();
@@ -109,7 +119,7 @@ describe('sdk-client object', () => {
 
   test('identify success withReasons', async () => {
     const carContext: LDContext = { kind: 'car', key: 'test-car' };
-    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, basicPlatform, {
+    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Enabled, mockPlatform, {
       logger,
       sendEvents: false,
       withReasons: true,
@@ -130,7 +140,7 @@ describe('sdk-client object', () => {
   test('identify success without auto env', async () => {
     defaultPutResponse['dev-test-flag'].value = false;
     const carContext: LDContext = { kind: 'car', key: 'test-car' };
-    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Disabled, basicPlatform, {
+    ldc = new LDClientImpl(testSdkKey, AutoEnvAttributes.Disabled, mockPlatform, {
       logger,
       sendEvents: false,
     });
@@ -148,6 +158,8 @@ describe('sdk-client object', () => {
   test('identify anonymous', async () => {
     defaultPutResponse['dev-test-flag'].value = false;
     const carContext: LDContext = { kind: 'car', anonymous: true, key: '' };
+
+    mockPlatform.crypto.randomUUID.mockReturnValue('random1');
 
     await ldc.identify(carContext);
     const c = ldc.getContext();
@@ -194,5 +206,47 @@ describe('sdk-client object', () => {
 
     expect(emitter.listenerCount('change')).toEqual(1);
     expect(emitter.listenerCount('error')).toEqual(1);
+  });
+
+  test('can complete identification using storage', async () => {
+    const data: Record<string, string> = {};
+    mockPlatform.storage.get.mockImplementation((key) => data[key]);
+    mockPlatform.storage.set.mockImplementation((key: string, value: string) => {
+      data[key] = value;
+    });
+    mockPlatform.storage.clear.mockImplementation((key: string) => {
+      delete data[key];
+    });
+
+    // First identify should populate storage.
+    await ldc.identify(context);
+
+    expect(logger.debug).not.toHaveBeenCalledWith('Identify completing with cached flags');
+
+    // Second identify should use storage.
+    await ldc.identify(context);
+
+    expect(logger.debug).toHaveBeenCalledWith('Identify completing with cached flags');
+  });
+
+  test('does not complete identify using storage when instructed to wait for the network response', async () => {
+    const data: Record<string, string> = {};
+    mockPlatform.storage.get.mockImplementation((key) => data[key]);
+    mockPlatform.storage.set.mockImplementation((key: string, value: string) => {
+      data[key] = value;
+    });
+    mockPlatform.storage.clear.mockImplementation((key: string) => {
+      delete data[key];
+    });
+
+    // First identify should populate storage.
+    await ldc.identify(context);
+
+    expect(logger.debug).not.toHaveBeenCalledWith('Identify completing with cached flags');
+
+    // Second identify would use storage, but we instruct it not to.
+    await ldc.identify(context, { waitForNetworkResults: true, timeout: 5 });
+
+    expect(logger.debug).not.toHaveBeenCalledWith('Identify completing with cached flags');
   });
 });