+++ packages/shared/sdk-client/src/polling/PollingProcessot.test.ts
@@ -29,7 +29,7 @@ function mockResponse(value: string, statusCode: number) {
 }
 
 /**
- * Mocks basicPlatform fetch. Returns the fetch jest.Mock object.
+ * Mocks fetch. Returns the fetch jest.Mock object.
  * @param remoteJson
  * @param statusCode
  */