+++ packages/shared/sdk-server/jest.config.js
@@ -3,6 +3,5 @@ module.exports = {
   testMatch: ['**/*.test.ts?(x)'],
   testEnvironment: 'node',
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-  collectCoverageFrom: ['src/**/*.ts'],
-  setupFilesAfterEnv: ['@launchdarkly/private-js-mocks/setup'],
+  collectCoverageFrom: ['src/**/*.ts']
 };